backend/pyproject.toml
[project]
name = "ziumsync"
version = "0.1.0"
description = "CDC Pipeline platform backed by Debezium"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.14.1",
    "celery>=5.4.0",
    "docker>=7.1.0",
    "fastapi>=0.115.8",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.1",
    "redis>=5.2.1",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ziumsync"]

[tool.ruff]
line-length = 120
target-version = "py310"
extend-exclude = ["alembic"]

[tool.ruff.lint]
select = ["E", "F", "I"] # PEP8 errors, pyflakes, isort
ignore = ["E501", "F403", "E711"]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
ignore_missing_imports = true
explicit_package_bases = true

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.12",
]
